requirements.txt
absl-py==2.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.10.0
astunparse==1.6.3
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
choreographer==1.0.10
click==8.3.1
cmdstanpy==1.3.0
colorama==0.4.6
contourpy==1.3.1
cryptography==46.0.5
cycler==0.12.1
decorator==5.1.1
deepface==0.0.93
distro==1.9.0
Django==5.2.7
django-cors-headers==4.9.0
django-extensions==4.1
django-sslserver==0.22
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dynamsoft_capture_vision_bundle==3.4.2000
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2026.2.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.65.0
google-pasta==0.2.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h5py==3.13.0
hf-xet==1.3.2
holidays==0.88
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.4.2
kaleido==1.0.0
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
logistro==1.1.0
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml_dtypes==0.5.1
moviepy==2.1.2
mpmath==1.3.0
mtcnn==1.0.0
namex==0.0.8
narwhals==2.2.0
networkx==3.6.1
numpy==2.1.3
openai==2.24.0
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
orjson==3.11.2
packaging==24.2
pandas==2.2.3
pillow==10.4.0
plotly==6.3.0
proglog==0.1.10
prophet==1.2.1
proto-plus==1.27.1
protobuf==5.29.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyasynchat==1.0.4
pyasyncore==1.0.4
PyAudio==0.2.14
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pyftpdlib==2.0.1
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==26.0.0
pyparsing==3.2.1
pyrxing==0.5.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.3
pyzbar==0.1.9
regex==2026.2.28
requests==2.32.3
retina-face==0.0.17
rich==13.9.4
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==76.1.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
stanio==0.5.1
sympy==1.14.0
tenacity==9.1.4
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.1
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
uritemplate==4.2.0
urllib3==2.3.0
websockets==16.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
